Refactor kiama to remove some abstraction
